pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "chintu"
version = "0.1.0"
description = "CHINTU: GDELT → TigerGraph narrative graph pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "python-dotenv>=1.0",
  "requests>=2.28",
]

[project.optional-dependencies]
tigergraph = ["pyTigerGraph>=2.0"]
api = [
  "flask>=3.0",
  "flask-cors>=4.0",
  "pyTigerGraph>=2.0",
  "httpx>=0.27",
  "openai>=1.40",
  "trafilatura>=1.8",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
